+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## [0.1.3] - 2021-xx-xx
 - Added file checksum to output JSON for canary mode
+- Added ability to check a previous version
+- Added option to save the file
 
 ## [0.1.2] - 2021-05-11
 - Fixing PGP key bug when JSON output file is not specified